improvement: svm.train in own module
